# docs/swagger.yaml
definitions:
  postgresdb.GetProductByIDRow:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
    type: object
  postgresdb.ListProductsRow:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
    type: object
  rest.BaseResponse:
    properties:
      error:
        type: string
      errorCode:
        type: integer
      success:
        type: boolean
    type: object
  rest.ProductRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      price:
        type: integer
      quantity:
        minimum: 0
        type: integer
    required:
    - description
    - name
    - price
    - quantity
    type: object
info:
  contact: {}
paths:
  /products:
    get:
      consumes:
      - application/json
      description: Get a list of all products in the warehouse
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/postgresdb.ListProductsRow'
              type: array
            type: object
        "500":
          description: List retrieval failed
          schema:
            $ref: '#/definitions/rest.BaseResponse'
      summary: List all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Add a new product to the warehouse
      parameters:
      - description: Product info
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/rest.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns ID of created product
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/rest.BaseResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/rest.BaseResponse'
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a product from the warehouse
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted
          schema:
            $ref: '#/definitions/rest.BaseResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/rest.BaseResponse'
        "500":
          description: Delete failed
          schema:
            $ref: '#/definitions/rest.BaseResponse'
      summary: Delete product by ID
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retrieve a single product from the warehouse
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product data
          schema:
            additionalProperties:
              $ref: '#/definitions/postgresdb.GetProductByIDRow'
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/rest.BaseResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/rest.BaseResponse'
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update product information in the warehouse
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product info
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/rest.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated
          schema:
            $ref: '#/definitions/rest.BaseResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/rest.BaseResponse'
        "500":
          description: Update failed
          schema:
            $ref: '#/definitions/rest.BaseResponse'
      summary: Update product by ID
      tags:
      - products
swagger: "2.0"